Dasch methods two_point, three_point, onion_peeling (note name change) file abel/dasch.py. Examples in one file examples/example_dasch_methods. Unit test test_dasch_methods.py. Updated docs. rm tools/basis.py. Methods called as abel.dasch.two_point_transform etc
