+++ abel/benchmark.py
@@ -9,10 +9,8 @@ import numpy as np
 from . import basex
 from . import hansenlaw
 from . import onion_bordas
-from . import onion_dasch
+from . import dasch
 from . import direct
-from . import two_point
-from . import three_point
 from . import tools
 
 class AbelTiming(object):
@@ -47,8 +45,8 @@ class AbelTiming(object):
                      'two_point_bs':    [],
                      'hansenlaw':      [],
                      'onion_bordas':  [],
-                     'onion_dasch': [],
-                     'onion_dasch_bs': [],
+                     'onion_peeling': [],
+                     'onion_peeling_bs': [],
                      'direct_python' : [] }
                      
         if direct.cython_ext:
@@ -90,16 +88,16 @@ class AbelTiming(object):
                 res_fabel['direct_python'].append(np.nan)
 
             t = time.time()
-            tpp=tools.basis.get_bs_cached("two_point", ni, basis_dir=None)
+            tpp=abel.dasch.get_bs_cached("two_point", ni, basis_dir=None)
             res_iabel['two_point_bs'].append((time.time()-t)*1000)
 
             t = time.time()
-            tp=tools.basis.get_bs_cached("three_point", ni, basis_dir=None)
+            tp=abel.dasch.get_bs_cached("three_point", ni, basis_dir=None)
             res_iabel['three_point_bs'].append((time.time()-t)*1000)
                 
             t = time.time()
-            od=tools.basis.get_bs_cached("onion_dasch", ni, basis_dir=None)
-            res_iabel['onion_dasch_bs'].append((time.time()-t)*1000)
+            od=abel.dasch.get_bs_cached("onion_peeling", ni, basis_dir=None)
+            res_iabel['onion_peeling_bs'].append((time.time()-t)*1000)
 
                 
             res_fabel['hansenlaw'].append(Timer(
@@ -109,13 +107,13 @@ class AbelTiming(object):
                 lambda: hansenlaw.hansenlaw_transform(x,  direction='inverse')).timeit(number=transform_repeat)*1000/transform_repeat)
 
             res_iabel['two_point'].append(Timer(
-                lambda: tools.basis.abel_transform(x, tpp)).timeit(number=transform_repeat)*1000/transform_repeat)
+                lambda: abel.dasch.abel_transform(x, tpp)).timeit(number=transform_repeat)*1000/transform_repeat)
                       
             res_iabel['three_point'].append(Timer(
-                lambda: tools.basis.core_transform(x, tp)).timeit(number=transform_repeat)*1000/transform_repeat)
+                lambda: abel.dasch.abel_transform(x, tp)).timeit(number=transform_repeat)*1000/transform_repeat)
                       
-            res_iabel['onion_dasch'].append(Timer(
-                lambda: tools.basis.core_transform(x, od)).timeit(number=transform_repeat)*1000/transform_repeat)
+            res_iabel['onion_peeling'].append(Timer(
+                lambda: abel.dasch.abel_transform(x, od)).timeit(number=transform_repeat)*1000/transform_repeat)
                       
             res_iabel['onion_bordas'].append(Timer(
                 lambda: onion_bordas.onion_bordas_transform(x)).timeit(number=transform_repeat)*1000/transform_repeat)
