+++ abel/tools/__init__.py
@@ -4,7 +4,6 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from . import analytical
-from . import basis
 from . import center
 from . import io
 from . import math
