+++ doc/example_dasch_methods.rst
@@ -1,0 +1,8 @@
+Example: Dasch methods
+======================
+
+
+.. plot:: ../examples/example_dasch_methods.py
+   :include-source:
+
+
