--- doc/example_onion_dasch.rst
@@ -1,8 +1,0 @@
-Example: onion_dasch
-=====================
-
-
-.. plot:: ../examples/example_onion_dasch.py
-   :include-source:
-
-
