+++ doc/examples.rst
@@ -12,3 +12,4 @@ Contents:
    example_basex_gaussian
    example_basex_photoelectron
    example_all_dribinski
+   example_dasch_methods
